+++ datasketches/src/bloom/mod.rs
@@ -126,8 +126,7 @@
 //! * Kirsch and Mitzenmacher (2008). "Less Hashing, Same Performance: Building a Better Bloom
 //!   Filter"
 
-mod builder;
 mod sketch;
 
-pub use self::builder::BloomFilterBuilder;
 pub use self::sketch::BloomFilter;
+pub use self::sketch::BloomFilterBuilder;
